pyproject.toml
[project]
name = "ai-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.120.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
